Display winner, award point and options at end
The winner is displayed so the user can see the result of the match.

A point is awarded to the winning player and displayed for the user.

Two options are given at the end of the match: a restart option that
reload the page and clear everything out, as well as a next round button
that will continue the current game.